pyproject.toml
[project]
name = "algdsnote-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
